Baseline: Fetch npm dependent packages (#2)
This PR adds a baseline to fetch downstream dependents of an npm package and determine their popularity according to their number of dependents and stars/forks on GitHub. In particular, the scaffold for a TypeScript/ts-node package is added, including a prototypical CLI (using oclif) and a first smoke-integration test (using jest). A CI script for GitHub Actions is added as well.