Baseline: Download npm dependent packages (#3)
This PR adds downloadDep() and a command for the CLI to download downstream dependent packages of an npm package (a550db8). Also, the oclif structure for the CLI is converted into a multi-command scaffold (ab2fcec). It is now possible to invoke the CLI using bin/run.js. Further internal changes include:

* getNpmDeps(): Add downloadGitHubData parameter (708c9e5)

* Refine and extend tests for getNpmDeps() (708c9e5)

* Test downloadDep() (32c875c)

* Revise test structure (see jest.config.ts, tsconfig.test.json, test/.jest, 32c875c)

* Fix .editorconfig for trailing whitespace in Markdown files (7c9fdfd)