Use type inference in package searcher (#33)
This PR introduces a more precise model for reference search that is based on the type analysis capabilities of the TypeScript compiler (src/references.ts). The existing PackageReferenceSearcher class is converted into a class hierarchy with two subclasses HeuristicPackageReferenceSearcher and TypePackageReferenceSearcher. Further changes are made to the general object model, including a new Package class and an extension of the existing Reference class, which previously was a type only and now also holds a FilePosition and a ReferenceKind indicating the kind of reference, i.e., usage, import, or occurrence. The public protocol on ReferenceSearcher is enabled to filter for relevant reference kinds.

The CLI command search is extended with additional flags to specify to search strategy to be used, filter for relevant reference kinds, and the type definitions for the package to be analyzed. Another flag for a simple aggregation of the results is added as well.

To test the new functionalities, the existing test structure is upgraded in order to support different search strategies and reference kinds (tests/references.test.ts). The expectedReferences fixtures are updated for the new properties (tests/references.test/). To complement the existing example users, example package definitions are provided for testing. Also, a short self-testing script is added to the CI in order to validate the given example packages and dependents (scripts/test-examples.sh).