Repackage 2: Split up project in single packages (#88)
This PR splits up the entire repository into a core package and a cli package. Most of the work in this PR focuses on establishing a connection, i.e. an npm file dependency between them and updating all paths such as for eslint and for the CI. Since this is not straightforward, a third example package is added as well which tests the simplest possible import case independently of the oclif burden in cli. Since yarn does not implement file dependencies properly in the non-berry version we are using (inter alia, it creates hard copies of the files required whereas npm uses transparent symlinks), for the cli and example package, we are switching back to npm, sigh. The core package gets a new facade for exports. Documentation for all packages is updated as well.